requirements.txt
PySide2
pillow
pyinstaller
ruff